Move BlossomServerListState to commons
